+++ code/src/nuvla/api/resources/credential.py
@@ -1,21 +1,13 @@
 
-from nuvla.api import Api as Nuvla
+from . import ResourceBase
 
 
-class Credential:
+class Credential(ResourceBase):
 
     resource = 'credential'
 
-    def __init__(self, nuvla: Nuvla):
-        self.nuvla = nuvla
-
-    def find(self, infra_service_id):
-        """
-        Returns list of credentials for `infra_service_id` infrastructure service.
-        :param infra_service_id: str, URI
-        :return: list
-        """
-        filters = "parent='{}'".format(infra_service_id)
+    def find_parent(self, parent):
+        filters = "parent='{}'".format(parent)
         creds = self.nuvla.search(self.resource, filter=filters, select="id")
         return creds.resources