+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+## [3.0.1] - 2020-03-27
+
+### Changed
+
+ - Add DeploymentParameter warapper resource,
+   fix Deployment credential_id method, and
+   Credential methods addition.
+
 ## [3.0.0] - 2020-03-25
 
 ### Changed